chore: Fix lint
We no longer allow importing from .test.ts files. This commit fixes all
the lint errors.
