+++ packages/replicache/src/filter-async-iterable.test.ts
@@ -1,7 +1,7 @@
 import {expect, test} from 'vitest';
 import {asyncIterableToArray} from './async-iterable-to-array.ts';
 import {filterAsyncIterable} from './filter-async-iterable.ts';
-import {makeAsyncIterable} from './merge-async-iterables.test.ts';
+import {makeAsyncIterable} from './make-async-iterable.ts';
 
 test('filterAsyncIterable', async () => {
   const t = async <V>(
