+++ packages/replicache/src/make-async-iterable.ts
@@ -1,0 +1,7 @@
+export async function* makeAsyncIterable<V>(
+  values: Iterable<V>,
+): AsyncIterable<V> {
+  for (const value of values) {
+    yield value;
+  }
+}
